requirements.txt
Django>=5.2.4
qrcode[pil]>=7.4.2
Pillow>=10.0.0
django-filter>=23.2
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
reportlab>=4.0.4
xlsxwriter>=3.1.2
python-decouple>=3.8
# Dependencias para MySQL
mysqlclient>=2.1.1
# Para manejar variables de entorno de forma segura
python-dotenv>=1.0.0
